ANDROID: GKI: add back pm_runtime_get_if_in_use()
In commit a55d55a30781 ("PM: runtime: Add pm_runtime_get_if_active()"),
pm_runtime_get_if_in_use() was replaced with an inline function.  That
would normally work just fine, except that function is part of the
Android abi, so the build breaks and so will any module that depends on
it.

So put the function back.

Bug: 161946584
Fixes: a55d55a30781 ("PM: runtime: Add pm_runtime_get_if_active()")
Change-Id: Ib7dd90758f1daca3d075d0bf0ae1bfc88b321218
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>